Pull Request

+++ app/Http/Controllers/MovieController.php
@@ -6,53 +6,58 @@ use Illuminate\Http\Request;
 
 class MovieController extends Controller
 {
-    /*FUNCION PARA EL USUARIO EL CUAL SE ENCARGA DE RECOGER TODAS LAS PELICULAS DE LA BASE DE DATOS*/
+    /*Se encarga de realizar mediante la variable MovieList la llamada a la tabla Movie de la base de datos y 
+    obtener todos los datos que esta contiene para mostrarlos posteriormente*/
     public function index(){
         $MovieList = Movie::all();
         return view('index', ['MovieList' => $MovieList]);
     }
 
-    /*FUNCION PARA EL USUARIOS EL CUAL SE ENCARGA DE MOSTRAR TODAS LAS PELICULAS Y DEVOLVER LA VISTA SHOW*/
+    /*Se encarga de realizar mediante la variable MovieList la llamada a la tabla Movie de la base de datos y 
+    obtener todos los datos que esta contiene para mostrarlos posteriormente y mostrarlo en la vista show*/
     public function show(){
         $movie = Movie::all();
         return view('Movies.show', ['movie' => $movie]);
     }
 
-    /*FUNCION DE MOSTRAR AL USUARIO PELICULAS SEPARADAS Y ORDENADAS DEPENDIENDO DEL ID DE LA PELICULA*/
+    /*Se encargar de mediante el id de la semana buscar todas las peliculas
+    en las cuales el id de la semana coincida con el seleccionado por el usurio
+    y de esta manera mostrar solo las pelícuals que pertenezcan a esa semana*/
     public function mostrar($semana_id){
         $movie = Movie::where('semana_id', $semana_id)->get();
         return view('Movies.mostrar', ['movies' => $movie, 'semana_id' => $semana_id]);
     }
 
-    /*FUNCION PARA EL USUARIO LA CUAL SE ENCARGA DE MOSTRAR TODOS LOS DATOS DE LA PELICULA OBTENIENDOLAS POR ID*/
+    /*Se encarga de mediante la llama al id de la pelicula buscar ese id y en caso de encontrarlo
+    mostrar toda la información de la película*/
     public function infoPeli($id){
         $movie = Movie::findOrFail($id);
-        return view('Movies.info', ['movie' => $movie]);
+        return view('Movies.mostrar', ['movie' => $movie]);
 
     }
 
     /*FUNCIONES ADMINISTADOR*/
 
-    /*FUNCION PARA EL ADMINISTADOR EN LA CUAL RECOGE TODA LA INFORMACION DE LAS PELICULAS*/
+    /*Se encarga de en la parte de administrador buscar mediante MovieList todas las películas*/
     public function ndex(){
         $MovieList = Movie::all();
         return view('admin_index', ['MovieList' => $MovieList]);
     }
 
-    /*FUNCION PARA EL ADMINISTADOR EN LA CUAL RECOGE TODAS LAS PELICULAS ALMACENADAS EN LA BASE DE DATOS Y MUESTRA LA VISTA SHOW*/
+    /*Se encarga de en la parte de administrador buscar mediante MovieList todas las películas y mostrarlas en la vista showAdmin*/
     public function showAdmin(){
         $movie = Movie::all();
         return view('Movies.showAdmin', ['movie' => $movie]);
     }
 
-    /*FUNCION PARA EL ADMINISTRADOR EL CUAL SE USA PARA CREAR PELICULAS Y DEVUELVE LA VISTA CREAR*/
+    /*Se encarga de crear nuevas películas*/
     public function create(){
         return view('movies.crear');
     }
 
-    /*FUNCION LA CUAL SIRVE PARA EL FORMULARIO PARA AÑADIR O MODIFICAR LAS PELICULAS Y QUE PIDE COMO OBLIGATORIO LOS DATOS QUE SE ENCUENTRAN GUARDADOS
-    EN LA BASE DE DATOS Y CREA UN NUEVO OBJETO MOVIE MEDIANTE LA VARIABLE MOVIE Y VALIDA LOS DATOS MEDIANTE LA VARIABLE validateData, LUEGO SE
-    ENCARGA DE GUARDAS LA VARIABLE Y NOS REDIRIGE A LA VISTA DE MOVIES DEL ADMINISTRADOR Y COGE LOS MENUS MEDIANTE LA ID*/
+    /*Se encarga de mediante el recibir datos de un request validar todos los datos de la base de datos
+    y de esta manera llamar al Movie de la base de datos y almacenar toda la informacion introducida en la variable movie,
+    para posteriormente guardarla*/
     public function store(Request $request){
         $validatedData = $request->validate([
             'title' => 'required',
@@ -60,6 +65,7 @@ class MovieController extends Controller
             'genero' => 'required',
             'duracion' => 'required', 
             'imagen' => 'required',
+            'trailer' => 'required',
             'director_id' => 'required',
             'actor_id' => 'required',
             'semana_id' => 'required'
@@ -75,21 +81,22 @@ class MovieController extends Controller
         $movie->actor_id = $validatedData['actor_id'];
         $movie->imagen = $validatedData['imagen'];
         $movie->semana_id = $validatedData['semana_id']; 
+        $movie->trailer = $validatedData['trailer'];
         $movie->fecha_lanzamiento = now(); 
 
         $movie->save();
         
-        return redirect()->route('movies.info', ['id' => $movie->id]);
+        return redirect()->route('admin.index', ['id' => $movie->id]);
     }
     
-    /*FUNCION PARA EL ADMINISTRADOR LA CUAL SIRVE PARA EDITAR LAS PELICULAS Y REDIRIGE AL FORMULARIO, ADEMAS DE BUSCAR LAS PELICULAS POR ID*/
+    /*Se encarga de buscar mediante id la película  y dar asi lugar a poder editarla en la vista from*/
     public function edit($id){
         $movie = Movie::find($id);
         return view('movies.form', ['movie' => $movie]);
     }
 
-    /*FUNCION LA CUAL SIRVE PARA ACTUALIZAR LA TABLA MOVIE Y SE ENCARGA DE BUSCAR MEDIANTE LA ID TODOS LOS DATOS DE LA TABLA DE MOVIE, GUARDARLOS Y MOSTRARLOS Y 
-    REIRIGE AL INDEX DEL AMDINISTRADOR*/
+    /*Se encarga de buscar mediante id la película seleccionada y de esta manera acceder a todos los datos actuales
+    y guardar los cambios realizados y mostrarlos*/
     public function update($id, Request $request){
         $movie = Movie::find($id);
         $movie->title = $request->title;
@@ -99,15 +106,21 @@ class MovieController extends Controller
         $movie->duracion = $request->duracion;
         $movie->fecha_lanzamiento = $request->fecha_lanzamiento;
         $movie->save();
-        return redirect()->route('admin.index');
+        return redirect()->route('Movies.showAdmin');
     }
     
 
-    /*FUNCION LA CUAL SIRVE PARA ELIMINAR MOVIES Y SE ENCARGA DE BUSCAR MEDIANTE EL ID EL MENU QUE QUEREMOS ELIMINAR*/
+    /*Se encarga de eliminar películas*/
     public function destroy($id){
         $movie = Movie::find($id);
-        $movie->delete();
-        return redirect()->route('admin.index');
-    }
+    
+        if ($movie) {
+            $movie->delete();
+            return redirect()->route('admin.moviesShow');
+        } else {
 
+            return redirect()->back()->withErrors(['error' => 'La película no se encontró']);
+        }
+
+}
 }
